pyproject.toml
[tool.poetry]
name = "wulkabot"
version = "0.1.0"
description = "Discord bot to help manage the Wulkanowy Discord server"
authors = ["JelNiSlaw <jelnislaw@gmail.com>"]
readme = "README.md"
packages = [{ include = "wulkabot" }]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3"
aiohttp = "^3.8"
python-dotenv = "^1.0"
beautifulsoup4 = "^4.12"

[tool.poetry.dev-dependencies]
black = "^23.3"
isort = "^5.12"
pyright = "^1.1"
pytest = "^7.3"

[tool.black]
verbose = true
line-length = 100
skip-magic-trailing-comma = true

[tool.isort]
profile = "black"
line_length = 100

[tool.pyright]
typeCheckingMode = "basic"
reportUnnecessaryTypeIgnoreComment = "warning"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
